chore: move DisableSyncPopup to own componentf
